+++ recipes/brocade.rb
@@ -20,4 +20,4 @@
 
 ['quantum', 'neutron'].include?(node['openstack']['compute']['network']['service_type']) || return
 
-include_recipe 'openstack-network::common'
+include_recipe 'openstack-network'